fix(macos): move hamburger to far-right of title bar; mirror right-anchored menus
On macOS the native traffic-light controls sit at the top-left, so the
hamburger (☰) menu — previously the first item in the tab strip — crowded
them. Move it to the far right of the window header (after the action
widgets), matching the Chrome top-right overflow-menu convention. Windows
and Linux keep it on the left, clear of their right-side caption buttons.

- Extract the menu into a self-contained <HamburgerMenu> component (every
  action was already a module-level store/RPC call — no TabBar coupling).
  Rendered in the tab strip on Win/Linux, in the header far-right on macOS.
- Add an opt-in `mirrored` mode to FlyoutMenu for right-anchored (left-
  opening) placement: items lay out right-to-left (icon right, label right-
  aligned), submenus prefer to open LEFT, and the disclosure chevron flips
  to point left — matching the direction the submenu actually opens (best
  practice for edge-anchored menus). Keyboard-shortcut hints are dropped in
  the mirrored menu. Scoped to the macOS far-right hamburger; every other
  menu keeps the standard left-to-right layout.
