+++ .changesets/1780494809-fix-macos-move-hamburger-menu-to-far-right-of-title-bar-mirr-dk4i.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+fix(macos): move hamburger menu to far-right of title bar; mirror right-anchored menus
